Tests reorganization
