+++ compiler-plugin/src/test/kotlin/arrow/meta/quotes/transform/plugins/TransformMetaPlugin.kt
@@ -1,0 +1,14 @@
+package arrow.meta.quotes.transform.plugins
+
+import arrow.meta.Meta
+import arrow.meta.Plugin
+import arrow.meta.phases.CompilerContext
+import kotlin.contracts.ExperimentalContracts
+
+open class TransformMetaPlugin : Meta {
+  @ExperimentalContracts
+  override fun intercept(ctx: CompilerContext): List<Plugin> = (
+    transformRemove
+      + transformReplace
+    )
+}
